refactor(consent): resolve a group's adder through one shared helper
